Fixed issue with not being able to retrieve audio file data
Turns out that files are sent to Downloads by default, which has its own URI. This meant external would be empty and then cause an Exception throw when trying to access any entries from the Cursor.

This has been fixed and it is now possible to retrieve Artist, Album and Song data from tracks that contain it.