Added Widgets to the "Now Playing" XML layouts
activity_now_playing.xml is responsible for the layout used by the "Now Playing" Activity. This contains a RecyclerView and some ImageButtons for playback control.

now_playing.xml is responsible for the UI widgets that will be used to display the metadata for the currently playing track. This will be used to inflate the RecyclerView inside of the activity_now_playing.xml layout.