Updated SongGridAdapter.java and NowPlayingActivity.java
Added onClickListeners for each ImageButton that is created. It creates a new instance of the Now Playing Activity and passes the Title, Album Name, Artist Name and Album ID for the new Activity to use.

NowPlayingActivity has also been updated to make use of the changes made in SongGridAdapter. So far only the details for the first ImageButton to be clicked can be retrieved - going back and clicking on another just presents the same data as the first.