Added NeedsCursor.java, Updated GridActivity.java and MediaStoreCursorLoader.java
Added the NeedsCursor interface for Activities (as long as they are AppCompatActivities) that need to use a Cursor object (this is for use of the Cursor created by using the MediaStore loader, but can be applied to any other purpose).

I have updated GridActivity to add a setCursor method, which is called by the MediaStoreCursorLoader class when the Cursor has been created.