Added MediaPlayerManager.java, Updated GridActivity.java and NowPlayingActivity.java
The MediaPlayerManager class has been created to be solely responsible for the control of the MediaPlayer instance and will itsefl be controlled by the NowPlayingScreen and other elements for controlling playback.

The NowPlayingActivity class has been updated to use this new class, and some methods have been added to remove code duplication.