Added functioning seekbar
Added an OnSeekBarChangeListener which calls MediaPlayerManager.seekTo() to set where in the track the playback should go to.

I have also added a ScheduledExecutorService that executes a Runnable object every 50 milliseconds. This Runnable object polls the MediaPlayerManager for the current position in the playback stream so that it can set the current position for the SeekBar.