Added onClickListener for album items in the GridView
In this onClickListener, the Intent for starting the AlbumActivity takes the album_id, album_cover, album_name, artist_name, first_year and num_songs values and passes them onto the instance of the activity.

This commit also includes the methods that have been created in order to get the data from the Cursor.